Change the way certificates are created
This keeps the old nasty behavior of having writes on read,
we should get rid of that some day.

Fixes #156
