+++ src/osha/oira/client/browser/webhelpers.py
@@ -1,15 +1,10 @@
 # coding=utf-8
-from datetime import date
 from euphorie.client.browser.webhelpers import WebHelpers
 from logging import getLogger
-from osha.oira.client.model import Certificate
 from osha.oira.client.model import UsersNotInterestedInCertificateStatusBox
-from plone import api
 from plone.memoize.instance import memoize
 from plone.namedfile.interfaces import INamedBlobImage
 from z3c.saconfig import Session
-
-import json
 
 
 log = getLogger(__name__)
@@ -36,43 +31,6 @@ class OSHAWebHelpers(WebHelpers):
             x, y = image.getImageSize()
             if x < 1000 or y < 430:
                 return True
-
-    def maybe_create_earned_certificate(self):
-        """Check if certificates are enabled in this country and if
-        the user has earned the certificate for this session.
-        In case the user needs a certificate, it will be created.
-        """
-        certificate_view = api.content.get_view(
-            "certificate", self.traversed_session, self.request
-        )
-        country = certificate_view.country
-
-        if not getattr(country, "certificates_enabled", False):
-            return
-
-        session = self.traversed_session.session
-        if session.completion_percentage < country.certificate_completion_threshold:
-            return
-
-        if (
-            Session.query(Certificate)
-            .filter(Certificate.session_id == session.session_id)
-            .first()
-        ):
-            return
-        Session.add(
-            Certificate(
-                session_id=session.session_id,
-                json=json.dumps({"date": date.today().strftime("%Y-%m-%d")}),
-            )
-        )
-
-    def update_completion_percentage(self, session=None):
-        completion_percentage = super(
-            OSHAWebHelpers, self
-        ).update_completion_percentage(session)
-        self.maybe_create_earned_certificate()
-        return completion_percentage
 
     def show_certificate_status_box(self):
         """Check if the current user should see his own certificate status box"""
